# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Energy and Water",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/mantenimiento_info": {
            "get": {
                "summary": "Obtener toda la información de mantenimiento",
                "operationId": "getAllMantenimientoInfo",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontraron datos"
                    }
                }
            },
            "post": {
                "summary": "Agregar nueva información de mantenimiento",
                "operationId": "addMantenimientoInfo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "texto",
                                    "activo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Información de mantenimiento creada con éxito"
                    },
                    "400": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/v1/mantenimiento_info/{id}": {
            "get": {
                "summary": "Obtener información de mantenimiento por ID",
                "operationId": "getMantenimientoInfoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la información de mantenimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    }
                }
            },
            "put": {
                "summary": "Actualizar o crear información de mantenimiento por ID",
                "operationId": "updateMantenimientoInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la información de mantenimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "texto",
                                    "activo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información de mantenimiento actualizada con éxito"
                    },
                    "201": {
                        "description": "Información de mantenimiento creada con éxito"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar información de mantenimiento por ID",
                "operationId": "deleteMantenimientoInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la información de mantenimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información de mantenimiento eliminada con éxito"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    }
                }
            },
            "patch": {
                "summary": "Actualizar parcialmente información de mantenimiento por ID",
                "operationId": "partialUpdateMantenimientoInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la información de mantenimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información de mantenimiento actualizada con éxito"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "400": {
                        "description": "No hay datos para actualizar"
                    }
                }
            }
        },
...
    }
}